test: add copy/paste independence tests
